+++ atm-finder/src/main/kotlin/net/progruzovik/alfabattle/atm/model/dto/AtmResponseDTO.kt
@@ -1,7 +1,7 @@
 package net.progruzovik.alfabattle.atm.model.dto
 
 data class AtmResponseDTO(
-    val deviceId: Long,
+    val deviceId: Int,
     val latitude: String?,
     val longitude: String?,
     val city: String,
